feat: Implement document type routing and dedicated views for screen and model documents
- Added `documentType` prop to `ChatPanel` for context-specific responses.
- Refactored `DocumentEditView` to route to `ScreenDocumentView` or `ModelDocumentView` based on document type.
- Created `ModelDocumentView` for editing model documents with dedicated sections.
- Created `ScreenDocumentView` for editing screen documents with dedicated sections.
- Updated `MarkdownEditor` to accept an optional placeholder prop.
- Enhanced chat functionality to provide context-aware responses based on document type.
