+++ src/components/Common/MarkdownEditor.tsx
@@ -6,11 +6,14 @@ import '@uiw/react-md-editor/markdown-editor.css';
 interface MarkdownEditorProps {
   value: string;
   onChange: (value: string) => void;
+  placeholder?: string;
 }
 
 export const MarkdownEditor: React.FC<MarkdownEditorProps> = React.memo(({ 
   value, 
-  onChange
+  onChange,
+  // @ts-ignore - placeholderは将来の拡張用に定義
+  placeholder = ''
 }) => {
   return (
     <div data-color-mode="light" data-testid="markdown-editor-container">
